services/fastapi_app/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
psycopg[binary]>=3.1
redis>=5.0
pytest>=8.0
httpx>=0.27
psycopg[binary]>=3.1
psycopg-pool>=3.2
python-dotenv